SignUp Functionality, password_reset confirm and complete
