+++ accounts/templates/accounts/password_reset_complete.html
@@ -1,0 +1,14 @@
+{% extends 'base_layout.html' %}
+{% block content %}
+    <div class="container my-3 p-3">
+        <div class="row justify-content-center">
+            <div class="col-lg-5">
+                <div class="card shadow-lg border-0 rounded-lg mt-0 mb-3">
+                    <div class="alert alert-info">
+                        Your password has been set. You may go ahead and <a href="{% url 'login' %}">Login Here</a>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+{% endblock content %}
