+++ accounts/templates/accounts/password_reset_confirm.html
@@ -1,0 +1,23 @@
+{% extends "base_layout.html" %}
+
+{% block title %}Enter new password{% endblock %}
+
+{% block content %}
+
+    {% if validlink %}
+
+        <h1>Set a new password!</h1>
+        <form method="POST">
+            {% csrf_token %}
+            {{ form.as_p }}
+            <input type="submit" value="Change my password">
+        </form>
+
+    {% else %}
+
+        <p>The password reset link was invalid, possibly because it has already been used. Please request a new password
+            reset.
+        </p>
+
+    {% endif %}
+{% endblock %}
