+++ accounts/urls.py
@@ -1,10 +1,19 @@
-from django.urls import path, include
-from . import views
+from django.contrib.auth import views as auth_views
+from django.urls import path
 
+from . import views
 
 app_name = 'accounts'
 
 urlpatterns = [
     path('', views.home_page, name='home'),
     path('login/', views.login_view, name='login'),
-    ]
+    path('signup/', views.signup_view, name='signup'),
+    path('reset/<uidb64>/<token>/',
+         auth_views.PasswordResetConfirmView.as_view(template_name='accounts/password_reset_confirm.html'),
+         name='password_reset_confirm'),
+    path('reset/done/',
+         auth_views.PasswordResetCompleteView.as_view(template_name='accounts/password_reset_complete.html'),
+         name='password_reset_complete.html'),
+    path('clear_cookie_consent_session/', views.clear_cookie_consent_session, name='clear_cookie_consent_session'),
+]
